DianShang/requirement.txt
amqp==2.2.2
billiard==3.5.0.3
celery==4.1.0
diff-match-patch==20121119
Django==2.0.3
django-crispy-forms==1.7.2
django-formtools==2.1
django-haystack==2.8.0
django-import-export==1.0.0
django-reversion==2.0.13
djangorestframework==3.8.2
et-xmlfile==1.0.1
future==0.16.0
httplib2==0.11.1
jdcal==1.3
jieba==0.39
kombu==4.1.0
nltk==3.2.5
numpy==1.14.2
odfpy==1.3.6
openpyxl==2.5.1
pdfkit==0.6.1
pillow==5.0.0
pyasn1==0.4.2
pycrypto==2.6.1
pymaging==0.1
pymaging-png==0.1
PyMySQL==0.8.0
python-redis-lock==3.2.0
pytz==2018.3
PyYAML==3.12
redis==2.10.6
redis-cache==0.1.3
redis-structures==0.1.6
rsa==3.4.2
scipy==1.0.1
six==1.11.0
tablib==0.12.1
unicodecsv==0.14.1
unknown==0.0.0
vine==1.1.4
vital-tools==0.1.8
wheel==0.24.0
Whoosh==2.7.4
xlrd==1.1.0
XlsxWriter==1.0.2
xlwt==1.3.0
rsa
